#Re-made all of concepts
- Ptr
: This object is based on smart pointer.
When an object is refered with this "pointer" it will managed auto-release.
If the object was pooled via MemoryPool then it will be free in MemoryPool when Unref was called.

- MemoryPool
: This object offers memories on heap memory.
MemoryPool is managing memories via paged memory.
But, Does Not Support merging fragments yet.
It will be updated later.

- RefCount
: The RefCount is just for counting over shared pointers.

- ArrayList
: The ArrayList is generated using MemoryPool.
The concept is based on ArrayList in common languages offering ArrayList.