> Bug Fix & Add LinkedList
!Bugs_

- MemoryPool
: After MemoryPool allocate some data block, it immediately release the block.
So, I move Ptr::Ref & Ptr::Unref into public definition space. And then, MemoryPool calls Ptr::Ref while allocating.

- Ptr
: When some pointer initialize with NULL Ptr::Copy ignore it.
That's the problem. Already initialized pointer never can be Null Pointer.
So, MemoryPool::Free is not called by Ptr.
I removed that ignoring codes and added it to Ref function.

>> LinkedList
: This will be performed using HTNode(Head & Tail Node).
It is not safe yet. Not recommanded to use it.